fix(docs): satisfy preview polish gates
