+++ docs/embed-providers.config.ts
@@ -1,0 +1,35 @@
+export function docsEmbedProviders(base: string) {
+  return {
+    pm: true,
+    googleMaps: true,
+    qiita: { fetch: false },
+    zenn: { fetch: false },
+    packageRegistry: { fetch: false },
+    playgrounds: { fetch: false },
+    vimeo: { fetch: false },
+    twitch: { fetch: false },
+    discord: true,
+    fediverse: true,
+    facebook: true,
+    threads: true,
+    instagram: true,
+    loom: true,
+    asciinema: true,
+    figma: true,
+    note: true,
+    googleSlides: true,
+    stackBlitz: true,
+    twitter: {
+      fetch: true,
+      mediaOutputDir: "public/twitter",
+      mediaPublicPath: `${base}twitter`,
+    },
+    bluesky: true,
+    webContainer: true,
+    spotify: true,
+    appleMusic: true,
+    speakerDeck: true,
+    audio: true,
+    video: true,
+  };
+}
