+++ npm/vite-plugin-ox-content/src/docs-config.test.ts
@@ -4,14 +4,16 @@ import { describe, expect, it } from "vite-plus/test";
 
 describe("official docs config", () => {
   it("enables authored containers for published component examples", async () => {
-    const config = await fs.readFile(
-      path.join(import.meta.dirname, "../../../docs/vite.config.ts"),
-      "utf8",
-    );
+    const configRoot = path.join(import.meta.dirname, "../../../docs");
+    const [config, embeds] = await Promise.all([
+      fs.readFile(path.join(configRoot, "vite.config.ts"), "utf8"),
+      fs.readFile(path.join(configRoot, "embed-providers.config.ts"), "utf8"),
+    ]);
 
     expect(config).toMatch(/containers:\s*true/);
-    expect(config).toMatch(/packageRegistry:\s*\{\s*fetch:\s*false\s*\}/);
-    expect(config).toMatch(/playgrounds:\s*\{\s*fetch:\s*false\s*\}/);
-    expect(config).toMatch(/googleMaps:\s*true/);
+    expect(config).toMatch(/embeds:\s*docsEmbedProviders\(base\)/);
+    expect(embeds).toMatch(/packageRegistry:\s*\{\s*fetch:\s*false\s*\}/);
+    expect(embeds).toMatch(/playgrounds:\s*\{\s*fetch:\s*false\s*\}/);
+    expect(embeds).toMatch(/googleMaps:\s*true/);
   });
 });
